Add files via upload
Hand detection detects hands in digital media and tracks these hands. There are applications in gesture recognition, sign language recognition, and virtual reality. Aspects of tracking that are also included in MediaPipe Holistic consist of face, hand, and body pose tracking when refined and measured at Real-time 3D. It allows for the simultaneous detection of all occurrences and key point extraction. It allows for interaction with technology using natural movements.